# change API to specify field assumptions at unpack callsite instead of globally