initial attempt at addressing #35. still needs: 

- [x] performance checks
- [x] deprecation warnings for the old `ImmutableStructType`/`MutableStructType`
- [x] documentation updates

I realized that there are more assumptions at play here than just ordering; I couldn't think of a succinct term that precisely captures them all, though, so I just use the vague `Exact{T}` - open to better names if folks have ideas! EDIT: @ararslan used the word "strict" here which I kind of like