@jrevels 2019-09-30
sooooooo this definitely results in a performance hit; I'm messing around with a type that looks like:

on master:
```
  julia> @time unpack(bytes, Vector{MyType});
    0.025953 seconds (32.24 k allocations: 2.767 MiB)

  julia> @time unpack(bytes, Vector{MutableMyType});
    0.034484 seconds (24.18 k allocations: 2.644 MiB)
```

this PR:

```
julia> @time unpack(bytes, Vector{MyType}; strict=(MyType,));
  0.032173 seconds (32.25 k allocations: 2.767 MiB)

julia> @time unpack(bytes, Vector{MyType});
  0.049074 seconds (48.35 k allocations: 3.996 MiB)
```

where `bytes` was generated via `pack(x::Vector{MyType})` where `length(x) == 8058` and `MyType` is defined like:

```
struct MyType
    a::Vector{Symbol}
    b::Symbol
    c::Float64
    d::String
    e::UInt64
    f::Symbol
    g::Union{Nothing,Dict{Symbol,Any}}
end
```

(and `MutableMyType` has the same fields, but is mutable).

The API here is so much nicer (IMO) that I'm tempted to accept the performance hit...I'd bet that some additional optimization can be done to further recover performance for at least the `strict` path, but for the non-`strict` path I think this PR's approach might just be fundamentally slower than the old approach (assuming both approaches were optimized as much as possible, which still probably isn't the case).