@jrevels 2019-10-02
Okay, did another round of benchmarking; [link to gist here](https://gist.github.com/jrevels/d2e9ac44d42c5fc7659281411131ff1e).

It turns out that the fast path overhead for this PR compared to master is actually negligible as long as the types being passed in are inferred as constants! This is great: essentially, there's no performance loss compared to master for the situation that master currently supports (globally assuming a given type should be parsed strictly).

The mutable path is still a bit slower, but I feel more comfortable with being able to recover that performance back later without too much API hassle (even if it just means reintroducing master's current implementation, for those folks that absolutely need that level of performance).

Now that I'm happy enough with the performance situation, I'll crank away on the docs/depwarns to get this merged.